+++ CHANGELOG.md
@@ -17,6 +17,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Removes multicolor feature
 - Removes ``pop`` method
 
+### Fixed
+- Fixes ``print`` method to mirror builtin
+
 [v1.0.8](https://github.com/jshwi/object_colors/releases/tag/v1.0.8)  - 2020-02-02
 ------------------------------------------------------------------------
 ### Added